modules/ClassificationModule/requirements.txt
torch==1.0
opencv-python==4.0.0.25
azure-iothub-device-client~=1.4.6
Pillow==6.0.0
